+++ spring-telqos-api/src/test/java/com/dwarfeng/springtelqos/api/integration/dubbo/DubboTestService.java
@@ -1,0 +1,12 @@
+package com.dwarfeng.springtelqos.api.integration.dubbo;
+
+/**
+ * Dubbo 测试使用的服务。
+ *
+ * @author DwArFeng
+ * @since 1.1.1
+ */
+public interface DubboTestService {
+
+    String echo(String content);
+}